[python] extract converter_compare.cpp from python_converter.cpp
Move 11 comparison/none/identity methods of python_converter into
src/python-frontend/converter/converter_compare.cpp:
resolve_comparison_operands_internal, has_unsupported_side_effects_internal,
compare_constants_internal, handle_indexed_comparison_internal,
handle_type_mismatches, handle_string_comparison, handle_none_comparison,
get_binary_operator_expr_for_is, get_negated_is_expr,
handle_string_type_mismatch, handle_type_identity_check.

Pure code movement — no logic change.
