+++ src/python-frontend/CMakeLists.txt
@@ -14,6 +14,7 @@ add_library(pythonfrontend STATIC
             python_language.cpp
             python_converter.cpp
             converter/converter_class.cpp
+            converter/converter_compare.cpp
             converter/converter_symbols.cpp
             converter/converter_types.cpp
             converter/converter_util.cpp
